requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
dj-database-url==2.2.0
django==5.1.2; python_version >= '3.10'
django-allauth==65.0.2; python_version >= '3.8'
django-cors-headers==4.4.0; python_version >= '3.8'
django-rest-auth==0.9.5
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
psycopg2==2.9.9; python_version >= '3.7'
pyjwt==2.9.0; python_version >= '3.8'
python-decouple==3.8
rest-framework-simplejwt==0.0.2; python_version >= '3.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
